color picker pt I
* added colour picker widget
* fixed popup widget with alignment
* tweaked geometry helper circle
* added vertex color to vec3 and vec4

+++ src/main/java/cinnamon/gui/widgets/WidgetList.java
@@ -91,7 +91,7 @@ public class WidgetList extends ContainerGrid {
         //update scrollbar
         updateList();
         scrollbar.setScrollAmount(3f / widgets.size()); //3 widgets per scroll
-        scrollbar.setHandleSize(scrollbar.getHeight() * getHeight() / getWidgetsHeight());
+        scrollbar.setHandleSize((int) (scrollbar.getHeight() * getHeight() / (float) getWidgetsHeight()));
 
         updateScrollbar();
     }
